Hexa Update

+++ Cursus/circle_1/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: antalvar <antalvar@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/06 15:51:38 by antalvar          #+#    #+#             */
-/*   Updated: 2024/05/08 01:38:15 by antalvar         ###   ########.fr       */
+/*   Updated: 2024/05/08 04:16:38 by antalvar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ int		ft_putchar(char c);
 int		ft_putdeci(int n);
 int		ft_putstr(char *s);
 int		ft_putunnbr(unsigned int i);
-int		ft_puthex(int n, char c);
-int		ft_puthex_format(int n, char c);
+int		ft_puthex(unsigned long n, char c);
+int		ft_puthex_format(unsigned long n, char c);
 int		ft_printf(const char *format, ...);
 int		ft_check_format(const char *str, va_list args);
 int		ft_putunnbr(unsigned int n);
